% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataManagement.R
\name{ICEScsv2xml}
\alias{ICEScsv2xml}
\title{Converts ICES csv to ICES XML}
\usage{
ICEScsv2xml(ices_csv = NULL)
}
\arguments{
\item{ices_csv}{The temperature in Fahrenheit.}
}
\description{
This function converts ices csv files into ices xml files
}
